Reject request lines containing NUL bytes
The code parsing the URI and HTTP version from the request line did
not work correctly when a NUL byte was included in the request line.
Since NUL bytes are not allowed in request lines, muhttpd now
rejects requests containing them with HTTP 400 Bad Request.

Thanks to Derek Abdine for reporting the problem.