Fix possible out-of-bounds read in configuration parsing
The newline detection in the configuration file parser looks at the
byte before the NUL terminator in the string. By creating a configuration
file where a NUL byte occurs as the first byte in a line, it was
possible to cause the configuration parser to read the byte before the
beginning of the string. This change avoids the out-of-bounds read.

Problem detected by afl-fuzz.